+++ src/jadeitite/entrypoint/sdl2.h
@@ -1,8 +1,0 @@
-#ifndef JADEITITE_SDL2_H
-#define JADEITITE_SDL2_H
-
-#include "jadeitite/log.h"
-#include "jadeitite/config.h"
-#include "jadeitite/renderer/sdl2.h"
-
-#endif //JADEITITE_SDL2_H